Restrukturierung Datenbankanfragen
