+++ src/main/java/de/prog3/projektarbeit/data/factories/PlayerFactory.java
@@ -1,29 +1,13 @@
 package de.prog3.projektarbeit.data.factories;
 
-import de.prog3.projektarbeit.data.database.JooqContextProvider;
 import de.prog3.projektarbeit.data.Position;
-import de.prog3.projektarbeit.data.jooq.tables.records.PlayerRecord;
 import de.prog3.projektarbeit.data.objects.Player;
-import de.prog3.projektarbeit.exceptions.PlayerNotFoundExeption;
 import de.prog3.projektarbeit.exceptions.ValidationException;
-import org.jooq.DSLContext;
-import org.jooq.Record;
-import org.jooq.Result;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.Optional;
-
-import static de.prog3.projektarbeit.data.jooq.tables.Player.PLAYER;
-import static de.prog3.projektarbeit.data.jooq.tables.Positions.POSITIONS;
 
 public class PlayerFactory {
-
-    private static final Logger logger = LoggerFactory.getLogger(PlayerFactory.class);
 
     private String firstName;
     private String lastName;
@@ -93,64 +77,5 @@ public class PlayerFactory {
     public Player build() throws ValidationException {
         validDateData(firstName, lastName, dateOfBirth, number, positions);
         return new Player(id, firstName, lastName, dateOfBirth, number, positions, teamId);
-    }
-
-    private static ArrayList<Position> getPlayerPositions(int playerId) {
-        DSLContext ctx = JooqContextProvider.getDSLContext();
-        Result<Record> positions = ctx.select().from(POSITIONS).where(POSITIONS.PLAYERID.eq(playerId)).fetch();
-        ArrayList<Position> playerPositions = new ArrayList<>();
-        if (positions.isEmpty()) {
-            return playerPositions;
-        }
-        for (Record r : positions) {
-            playerPositions.add(Position.valueOf(r.get(POSITIONS.POSITION)));
-        }
-        return playerPositions;
-    }
-
-    public static Optional<Player> extractPlayerFromRecord(Record record) {
-        Player player = null;
-        if (record.get(PLAYER.ID) != null) {
-            try {
-                player = new Player(record.get(PLAYER.ID), record.get(PLAYER.FIRSTNAME), record.get(PLAYER.LASTNAME), Player.parseStringToDate(record.get(PLAYER.DATEOFBIRTH)), record.get(PLAYER.NUMBER), getPlayerPositions(record.get(PLAYER.ID)), record.get(PLAYER.TEAM_ID));
-            } catch (ParseException e) {
-                logger.error("Fehler beim Parsen des Geburtsdatums: ", e);
-            }
-        }
-        return Optional.ofNullable(player);
-    }
-
-    public static Player getPlayerById(int id) throws PlayerNotFoundExeption {
-        DSLContext ctx = JooqContextProvider.getDSLContext();
-        PlayerRecord record = (PlayerRecord) ctx.select().from(PLAYER).where(PLAYER.ID.eq(id)).fetchOne();
-        Player player = null;
-        try {
-            if (record == null) {
-                throw new PlayerNotFoundExeption("Spieler mit der ID " + id + " nicht gefunden");
-            }
-            Optional<Integer> teamId = Optional.ofNullable(record.getTeamId());
-            player = new Player(record.getId(), record.getFirstname(), record.getLastname(), Player.parseStringToDate(record.getDateofbirth()), record.getNumber(), getPlayerPositions(record.getId()), teamId.orElse(0));
-        } catch (ParseException e) {
-            throw new PlayerNotFoundExeption("Spieler mit der ID " + id + " nicht gefunden");
-        }
-        return player;
-    }
-
-    public static HashMap<Integer, Player> getFreeAgents() {
-        DSLContext ctx = JooqContextProvider.getDSLContext();
-        Result<Record> result = ctx.select().from(PLAYER).where(PLAYER.TEAM_ID.eq(0).or(PLAYER.TEAM_ID.isNull())).fetch();
-        HashMap<Integer, Player> players = new HashMap<>();
-        for (Record r : result) {
-            PlayerRecord record = (PlayerRecord) r;
-            try {
-                Optional<Integer> teamId = Optional.ofNullable(record.getTeamId());
-                Player player = new Player(record.getId(), record.getFirstname(), record.getLastname(), Player.parseStringToDate(record.getDateofbirth()), record.getNumber(), getPlayerPositions(((PlayerRecord) r).getId()), teamId.orElse(0));
-                players.put(player.getId(), player);
-            } catch (ParseException e) {
-                logger.error("Fehler beim Parsen des Geburtsdatums: ", e);
-                throw new RuntimeException(e);
-            }
-        }
-        return players;
     }
 }
