+++ src/main/java/de/prog3/projektarbeit/data/objects/Player.java
@@ -1,9 +1,6 @@
 package de.prog3.projektarbeit.data.objects;
 
-import de.prog3.projektarbeit.data.database.JooqContextProvider;
 import de.prog3.projektarbeit.data.Position;
-import de.prog3.projektarbeit.exceptions.UnableToSavePlayerExeption;
-import org.jooq.DSLContext;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -86,55 +83,6 @@ public class Player extends DataObject {
     public static String parseDateToString(Date date) throws ParseException {
         SimpleDateFormat format = new SimpleDateFormat("dd-MM-yyyy");
         return format.format(date);
-    }
-
-    public void save() throws UnableToSavePlayerExeption {
-        try {
-            DSLContext ctx = JooqContextProvider.getDSLContext();
-            if(super.getId() == 0){
-                ctx.insertInto(PLAYER)
-                        .columns(PLAYER.FIRSTNAME, PLAYER.LASTNAME, PLAYER.DATEOFBIRTH, PLAYER.NUMBER)
-                        .values(this.firstName, this.lastName, parseDateToString(this.dateOfBirth), this.number)
-                        .onDuplicateKeyUpdate()
-                        .set(PLAYER.FIRSTNAME, this.firstName)
-                        .set(PLAYER.LASTNAME, this.lastName)
-                        .set(PLAYER.DATEOFBIRTH, parseDateToString(dateOfBirth))
-                        .set(PLAYER.NUMBER, this.number)
-                        .set(PLAYER.TEAM_ID, this.teamId)
-                        .execute();
-                ctx.select(PLAYER.ID).from(PLAYER).where(PLAYER.FIRSTNAME.eq(this.firstName)).and(PLAYER.LASTNAME.eq(this.lastName)).fetch().forEach(record -> super.setId(record.get(PLAYER.ID)));
-            } else {
-                ctx.update(PLAYER)
-                        .set(PLAYER.FIRSTNAME, this.firstName)
-                        .set(PLAYER.LASTNAME, this.lastName)
-                        .set(PLAYER.DATEOFBIRTH, parseDateToString(dateOfBirth))
-                        .set(PLAYER.NUMBER, this.number)
-                        .set(PLAYER.TEAM_ID, this.teamId)
-                        .where(PLAYER.ID.eq(super.getId()))
-                        .execute();
-            }
-            ArrayList<Position> oldPositions = new ArrayList<>();
-            ctx.select().from(POSITIONS).where(POSITIONS.PLAYERID.eq(super.getId())).fetch().forEach(record -> oldPositions.add(Position.valueOf(record.get(POSITIONS.POSITION))));
-            ArrayList<Position> added = new ArrayList<>(positions);
-            ArrayList<Position> removed = new ArrayList<>(oldPositions);
-            added.removeAll(removed);
-            removed.removeAll(positions);
-
-            removed.forEach(position ->
-                    ctx.deleteFrom(POSITIONS)
-                            .where(POSITIONS.PLAYERID.eq(super.getId())).and(POSITIONS.POSITION.eq(position.name()))
-                            .execute()
-            );
-
-            added.forEach(position ->
-                    ctx.insertInto(POSITIONS)
-                            .columns(POSITIONS.PLAYERID, POSITIONS.POSITION)
-                            .values(super.getId(), position.name())
-                            .execute()
-            );
-        } catch (ParseException e){
-            throw new UnableToSavePlayerExeption("Player {" +this.firstName + " " + this.lastName + "} could not be saved");
-        }
     }
 
     public String getFullName(){
